Add pretty_format helper
