+++ crates/macros/cgp-macro-core/Cargo.toml
@@ -9,7 +9,8 @@ rust-version = { workspace = true }
 keywords     = { workspace = true }
 
 [dependencies]
-syn             = { version = "2.0.95", features = [ "full", "extra-traits", "visit", "visit-mut" ] }
-quote           = "1.0.38"
-proc-macro2     = "1.0.92"
-itertools       = "0.14.0"
+syn                 = { version = "2.0.95", features = [ "full", "extra-traits", "visit", "visit-mut" ] }
+quote               = { version = "1.0.38" }
+proc-macro2         = { version = "1.0.92" }
+itertools           = { version = "0.14.0" }
+prettyplease        = { version = "0.2.37" }
