+++ crates/macros/cgp-macro-core/src/functions/mod.rs
@@ -1,19 +1,23 @@
 mod camel_case;
 mod delegated_impls;
 mod field;
+mod format;
 mod generics;
 mod getter;
 mod implicits;
 mod is_provider_params;
 mod parse_internal;
 mod snake_case;
+mod strip;
 
 pub use camel_case::*;
 pub use delegated_impls::*;
 pub use field::*;
+pub use format::*;
 pub use generics::*;
 pub use getter::*;
 pub use implicits::*;
 pub use is_provider_params::*;
 pub use parse_internal::*;
 pub use snake_case::*;
+pub use strip::*;
